backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
mysql-connector-python
pymilvus
langchain
openai
requests
python-dotenv
pytest
motor
pymongo
passlib
